Move create Tweet to useTweets
